Move WebhookUrl component to jobs-form-fields.tsx
Keeps job settings field components together in one file, consistent
with JobWorktreeOption, JobFullAccessOption, and JobKeepAgentOption.
